change id type in  BooksOrderr and fixing method in BooksOrderrController.

+++ src/main/java/restapi/webapp/OrderrNotFoundException.java
@@ -1,7 +1,7 @@
 package restapi.webapp;
 
 public class OrderrNotFoundException extends RuntimeException{
-    public OrderrNotFoundException(int numberOfOrder){
+    public OrderrNotFoundException(Long numberOfOrder){
         super("There is no order corresponding to numberOfOrder = " + numberOfOrder);
     }
 }
